# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: P2P Game Item Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/topup:
    post:
      operationId: topupWallet
      tags: [wallet]
      summary: Top up wallet balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupRequest"
      responses:
        "200":
          description: Top up success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /wallet/withdraw:
    post:
      operationId: requestWithdrawal
      tags: [wallet]
      summary: Request a withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequest"
      responses:
        "201":
          description: Withdrawal request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    WalletBalance:
      type: object
      properties:
        balance:
          type: number
          format: float
        escrowBalance:
          type: number
          format: float
      required:
        - balance
        - escrowBalance
...
    TopupRequest:
      type: object
      properties:
        amount:
          type: number
          format: float
        method:
          type: string
          enum: [bank_transfer, qris, gopay, ovo, dana]
      required:
        - amount
        - method
...
    WithdrawalRequest:
      type: object
      properties:
        amount:
          type: number
          format: float
        method:
          type: string
        accountNumber:
          type: string
        accountName:
          type: string
      required:
        - amount
        - method
        - accountNumber
        - accountName
...
    Withdrawal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
          format: float
        method:
          type: string
        accountNumber:
          type: string
        accountName:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        adminNote:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - amount
        - method
        - accountNumber
        - accountName
        - status
        - createdAt
        - updatedAt
...